set and made available color palette to app
