+++ client/package.json
@@ -11,7 +11,8 @@
   },
   "devDependencies": {
     "@material-ui/core": "^4.7.0",
-    "@material-ui/icons": "^4.5.1"
+    "@material-ui/icons": "^4.5.1",
+    "@material-ui/styles": "^4.6.0"
   },
   "scripts": {
     "start": "react-scripts start",
