+++ client/src/App.js
@@ -3,9 +3,28 @@ import { Route, Switch, BrowserRouter as Router } from 'react-router-dom';
 import './App.css';
 import Login from './pages/Login';
 import Dashboard from './pages/Dashboard';
+// create theme for entire project
+import { createMuiTheme, ThemeProvider } from '@material-ui/core/styles';
+
+const theme = createMuiTheme({
+  palette: {
+    primary: {
+      light: '#3f3a3a',
+      main: '#191414',
+      dark: '#000000',
+      contrastText: '#fff',
+    },
+    secondary: {
+      light: '#62ec83',
+      main: '#1db954',
+      dark: '#008827',
+      contrastText: '#fff',
+    },
+  },
+});
 
 export default function App(props) {
-  
+
     const App = () => (
       <div>
         <Switch>
@@ -19,11 +38,13 @@ export default function App(props) {
       </div>
     )
     return (
-      <Router>
-        <Switch>
-          <App/>
-        </Switch>
-      </Router>
+      <ThemeProvider theme={theme}>
+        <Router>
+          <Switch>
+            <App/>
+          </Switch>
+        </Router>
+      </ThemeProvider>
     );
   }
