+++ client/src/hooks/useDashboardData.js
@@ -32,33 +32,33 @@ export default function useDashboardData() {
       }).catch((e) => console.log('error:', e))
   }, []);
 
-  // useEffect(() => {
-  //   if (state.token) {
-  //     getPerformers().then(events => {
-  //       console.log("test", events);
-  //       setState(prev => ({ ...prev, events }));
-  //     });
-  //   }
-  // }, [state.token]);
+  useEffect(() => {
+    if (state.token) {
+      getPerformers().then(events => {
+        console.log("test", events);
+        setState(prev => ({ ...prev, events }));
+      });
+    }
+  }, [state.token]);
 
-  // useEffect(() => {
-  //   if (state.token && state.events && state.events !== {}) {
-  //     getArtists(state.token, state.events)
-  //       .then(artists => {
-  //         setState(prev => ({ ...prev, artists }));
-  //     });
-  //   }
-  // }, [state.token, state.events]);
+  useEffect(() => {
+    if (state.token && state.events && state.events !== {}) {
+      getArtists(state.token, state.events)
+        .then(artists => {
+          setState(prev => ({ ...prev, artists }));
+      });
+    }
+  }, [state.token, state.events]);
 
 
-  // useEffect(() => {
-  //   if (state.token) {
-  //     getSongs(state.token, state.artists)
-  //       .then(songs => {
-  //         setState(prev => ({...prev, songs}))
-  //       })
-  //   }
-  // }, [state.token, state.events, state.artists]);
+  useEffect(() => {
+    if (state.token) {
+      getSongs(state.token, state.artists)
+        .then(songs => {
+          setState(prev => ({...prev, songs}))
+        })
+    }
+  }, [state.token, state.events, state.artists]);
 
   // On Mount, load Spotify Web Playback SDK script
   useEffect(() => {
